+++ edge.h
@@ -4,12 +4,12 @@ struct Node;
 struct Edge
 {
     int dstNodeId;
-    double weight;
-    double *transProbTable;
+    float weight;
+    float *transProbTable;
     int *aliasTable;
     Edge *nextEdge;
 
-    Edge(int dst, int weight);
+    Edge(int dst, float weight);
     void edgePreprocess(int src, Graph* G);
     bool hasEdge(Node *vertex, int src, int dst);
 };
